+++ src/main/java/racingcar/Application.java
@@ -1,7 +1,11 @@
 package racingcar;
 
+import racingcar.game.Game;
+
 public class Application {
     public static void main(String[] args) {
         // TODO: 프로그램 구현
+        Game game = new Game();
+        game.start();
     }
 }
