+++ src/main/java/racingcar/car/CarFactory.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.car;
 
 import java.util.ArrayList;
 import java.util.List;
