+++ src/main/java/racingcar/constant/RacingGameMessage.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.constant;
 
 public class RacingGameMessage {
     public static final String INPUT_CAR_NAMES_PROMPT = "경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)";
