+++ src/main/java/racingcar/io/PrintHandler.java
@@ -1,16 +1,12 @@
-package racingcar;
+package racingcar.io;
+
+import racingcar.car.Car;
 
 import java.util.List;
 
-import static racingcar.RacingGameMessage.*;
+import static racingcar.constant.RacingGameMessage.*;
 
 public class PrintHandler {
-//    public class RacingGameMessage {
-//        public static final String INPUT_CAR_NAMES = "경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)";
-//        public static final String INPUT_TRIAL_NUMBER = "시도할 회수는 몇회인가요?";
-//        public static final String RESULT_RUNNING = "실행 결과";
-//        public static final String FINAL_WINNER = "최종 우승자 : ";
-//    }
     public static void inputCarNamesPrompt() {
         System.out.println(INPUT_CAR_NAMES_PROMPT);
     }
@@ -42,7 +38,7 @@ public class PrintHandler {
 
     private static void printDash(int location) {
         for (int i = 0; i < location; i++) {
-            System.out.print("-");
+            System.out.print(RESULT_LOCATION_EXPRESSION);
         }
     }
 }
