+++ src/main/java/racingcar/util/StringManipulator.java
@@ -1,12 +1,12 @@
-package racingcar;
+package racingcar.util;
 
 import java.util.Arrays;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
-import static racingcar.Constants.MAX_CAR_NAME_LEN;
-import static racingcar.Constants.MIN_CAR_NAME_LEN;
+import static racingcar.constant.Constants.MAX_CAR_NAME_LEN;
+import static racingcar.constant.Constants.MIN_CAR_NAME_LEN;
 
 public class StringManipulator {
     public static String[] splitSeparator(String sentence, String separator) {
